backend/requirements.txt
flask
scikit-learn
pandas
python-docx
pdfminer.six
nltk